fixed image centering and zooming type

+++ c_implementation/mandelbrotTools.c
@@ -45,7 +45,7 @@ int isInMandelbrotSet(Complex* c, int processPower, int setPoint) {
 }
 
 void createMandelbrot(unsigned char* pixelBuffer, int width, int height,
-    int processPower, int setPoint, int centerReal, int centerImag, int zoom)
+    int processPower, int setPoint, double centerReal, double centerImag, double zoom)
     {
     if (pixelBuffer == NULL || width <= 0 || height <= 0) {
         return;
